refactor(rust-api): extract timeline + person handlers into modules (sc-1636)
Move the timeline endpoint handlers (list/create/get/update timeline,
create_timeline_export, extract_timeline_frame) into timelines.rs and
the person-track handlers (list/get tracks, save_corrections,
create_person_detection_job, create_person_track_job) into person.rs.

Same pattern as auth/characters: `use super::*;` + pub(crate) handlers
glob-re-imported at the root so router wiring stays unqualified.
Behavior-preserving. lib.rs 7,890 -> 7,639 lines.
