+++ apps/rust-api/src/lib.rs
@@ -66,6 +66,10 @@ mod auth;
 use auth::{access_control, cors_layer, is_authorized};
 mod characters;
 use characters::*;
+mod timelines;
+use timelines::*;
+mod person;
+use person::*;
 
 const PUBLIC_PATHS: &[&str] = &[
     "/api/v1/health",
@@ -2288,257 +2292,6 @@ fn parse_single_byte_range(value: &str, total: u64) -> Option<(u64, u64)> {
         return None;
     }
     Some((start, end))
-}
-
-async fn list_timelines(
-    State(state): State<AppState>,
-    Path(project_id): Path<String>,
-) -> Result<Json<Vec<sceneworks_core::project_store::TimelineSummary>>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| store.list_timelines(&project_id)).await?,
-    ))
-}
-
-async fn create_timeline(
-    State(state): State<AppState>,
-    Path(project_id): Path<String>,
-    ApiJson(payload): ApiJson<TimelineCreateRequest>,
-) -> Result<(StatusCode, Json<Value>), ApiError> {
-    let timeline = project_call(state, move |store| {
-        store.create_timeline(
-            &project_id,
-            &payload.name,
-            &payload.aspect_ratio,
-            payload.fps,
-        )
-    })
-    .await?;
-    Ok((StatusCode::CREATED, Json(timeline)))
-}
-
-async fn get_timeline(
-    State(state): State<AppState>,
-    Path((project_id, timeline_id)): Path<(String, String)>,
-) -> Result<Json<Value>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| {
-            store.get_timeline(&project_id, &timeline_id)
-        })
-        .await?,
-    ))
-}
-
-async fn update_timeline(
-    State(state): State<AppState>,
-    Path((project_id, timeline_id)): Path<(String, String)>,
-    ApiJson(payload): ApiJson<TimelineSaveRequest>,
-) -> Result<Json<Value>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| {
-            store.save_existing_timeline(&project_id, &timeline_id, payload.timeline)
-        })
-        .await?,
-    ))
-}
-
-async fn create_timeline_export(
-    State(state): State<AppState>,
-    Path((project_id, timeline_id)): Path<(String, String)>,
-    ApiJson(payload): ApiJson<TimelineExportRequest>,
-) -> Result<(StatusCode, Json<JobSnapshot>), ApiError> {
-    validate_timeline_export(&payload)?;
-    let timeline_result = project_call(state.clone(), {
-        let project_id = project_id.clone();
-        let timeline_id = timeline_id.clone();
-        move |store| store.timeline_file_and_document(&project_id, &timeline_id)
-    })
-    .await?;
-    let timeline_name = timeline_result
-        .document
-        .get("name")
-        .and_then(Value::as_str)
-        .unwrap_or("Timeline")
-        .to_owned();
-    let mut job_payload = JsonObject::new();
-    job_payload.insert("projectId".to_owned(), Value::String(project_id.clone()));
-    job_payload.insert("timelineId".to_owned(), Value::String(timeline_id));
-    job_payload.insert("timelineName".to_owned(), Value::String(timeline_name));
-    job_payload.insert(
-        "timelinePath".to_owned(),
-        Value::String(timeline_result.file.relative_path),
-    );
-    job_payload.insert("resolution".to_owned(), json!(payload.resolution));
-    job_payload.insert("fps".to_owned(), json!(payload.fps));
-    let job = create_generation_job(
-        state,
-        JobType::TimelineExport,
-        Some(project_id),
-        None,
-        job_payload,
-        payload.requested_gpu,
-    )
-    .await?;
-    Ok((StatusCode::CREATED, Json(job)))
-}
-
-async fn extract_timeline_frame(
-    State(state): State<AppState>,
-    Path((project_id, timeline_id, item_id)): Path<(String, String, String)>,
-    ApiJson(payload): ApiJson<FrameExtractRequest>,
-) -> Result<(StatusCode, Json<JobSnapshot>), ApiError> {
-    validate_frame_extract(&payload)?;
-    let timeline_result = project_call(state.clone(), {
-        let project_id = project_id.clone();
-        let timeline_id = timeline_id.clone();
-        move |store| store.timeline_file_and_document(&project_id, &timeline_id)
-    })
-    .await?;
-    let item = find_timeline_item(&timeline_result.document, &item_id)?;
-    let source_asset_id = required_string_field(item, "assetId")?.to_owned();
-    let timestamp = source_timestamp_for_item(item, payload.playhead_seconds)?;
-    let timeline_name = timeline_result
-        .document
-        .get("name")
-        .and_then(Value::as_str)
-        .unwrap_or("Timeline")
-        .to_owned();
-    let mut job_payload = JsonObject::new();
-    job_payload.insert("projectId".to_owned(), Value::String(project_id.clone()));
-    job_payload.insert("timelineId".to_owned(), Value::String(timeline_id));
-    job_payload.insert("timelineName".to_owned(), Value::String(timeline_name));
-    job_payload.insert(
-        "timelinePath".to_owned(),
-        Value::String(timeline_result.file.relative_path),
-    );
-    job_payload.insert("timelineItemId".to_owned(), Value::String(item_id));
-    job_payload.insert("sourceAssetId".to_owned(), Value::String(source_asset_id));
-    job_payload.insert("sourceTimestamp".to_owned(), json!(timestamp));
-    job_payload.insert(
-        "playheadSeconds".to_owned(),
-        json!(payload.playhead_seconds),
-    );
-    job_payload.insert(
-        "intendedUse".to_owned(),
-        Value::String(payload.intended_use),
-    );
-    let job = create_generation_job(
-        state,
-        JobType::FrameExtract,
-        Some(project_id),
-        None,
-        job_payload,
-        payload.requested_gpu,
-    )
-    .await?;
-    Ok((StatusCode::CREATED, Json(job)))
-}
-
-async fn list_person_tracks(
-    State(state): State<AppState>,
-    Path(project_id): Path<String>,
-) -> Result<Json<Vec<Value>>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| store.list_person_tracks(&project_id)).await?,
-    ))
-}
-
-async fn get_person_track(
-    State(state): State<AppState>,
-    Path((project_id, track_id)): Path<(String, String)>,
-) -> Result<Json<Value>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| {
-            store.get_person_track(&project_id, &track_id)
-        })
-        .await?,
-    ))
-}
-
-async fn save_person_track_corrections(
-    State(state): State<AppState>,
-    Path((project_id, track_id)): Path<(String, String)>,
-    ApiJson(payload): ApiJson<PersonTrackCorrectionsRequest>,
-) -> Result<Json<Value>, ApiError> {
-    Ok(Json(
-        project_call(state, move |store| {
-            store.set_person_track_corrections(&project_id, &track_id, payload.corrections)
-        })
-        .await?,
-    ))
-}
-
-async fn create_person_detection_job(
-    State(state): State<AppState>,
-    Path(project_id): Path<String>,
-    ApiJson(payload): ApiJson<PersonDetectionJobRequest>,
-) -> Result<(StatusCode, Json<JobSnapshot>), ApiError> {
-    validate_person_detection_job(&payload)?;
-    let project_name = project_call(state.clone(), {
-        let project_id = project_id.clone();
-        move |store| store.project_stem(&project_id)
-    })
-    .await?;
-    let mut job_payload = JsonObject::new();
-    job_payload.insert("projectId".to_owned(), Value::String(project_id.clone()));
-    job_payload.insert(
-        "sourceAssetId".to_owned(),
-        Value::String(payload.source_asset_id),
-    );
-    job_payload.insert(
-        "sourceTimestamp".to_owned(),
-        payload.source_timestamp.map_or(Value::Null, Value::from),
-    );
-    if payload.preview {
-        job_payload.insert("preview".to_owned(), Value::Bool(true));
-    }
-    let job = create_generation_job(
-        state,
-        JobType::PersonDetect,
-        Some(project_id),
-        Some(project_name),
-        job_payload,
-        payload.requested_gpu,
-    )
-    .await?;
-    Ok((StatusCode::CREATED, Json(job)))
-}
-
-async fn create_person_track_job(
-    State(state): State<AppState>,
-    Path(project_id): Path<String>,
-    ApiJson(payload): ApiJson<PersonTrackJobRequest>,
-) -> Result<(StatusCode, Json<JobSnapshot>), ApiError> {
-    validate_person_track_job(&payload)?;
-    let project_name = project_call(state.clone(), {
-        let project_id = project_id.clone();
-        move |store| store.project_stem(&project_id)
-    })
-    .await?;
-    let mut job_payload = JsonObject::new();
-    job_payload.insert("projectId".to_owned(), Value::String(project_id.clone()));
-    job_payload.insert(
-        "sourceAssetId".to_owned(),
-        Value::String(payload.source_asset_id),
-    );
-    job_payload.insert(
-        "representativeFrameAssetId".to_owned(),
-        Value::String(payload.representative_frame_asset_id),
-    );
-    job_payload.insert("detection".to_owned(), Value::Object(payload.detection));
-    job_payload.insert("trackName".to_owned(), Value::String(payload.track_name));
-    if payload.preview {
-        job_payload.insert("preview".to_owned(), Value::Bool(true));
-    }
-    let job = create_generation_job(
-        state,
-        JobType::PersonTrack,
-        Some(project_id),
-        Some(project_name),
-        job_payload,
-        payload.requested_gpu,
-    )
-    .await?;
-    Ok((StatusCode::CREATED, Json(job)))
 }
 
 async fn create_image_job(
